refactor: Merging info and migration commands
