+++ packages/checkup-plugin-ember/package.json
@@ -40,8 +40,7 @@
       "@oclif/plugin-help"
     ],
     "hooks": {
-      "register-info-tasks": "./lib/hooks/register-info-tasks",
-      "register-migration-tasks": "./lib/hooks/register-migration-tasks",
+      "register-tasks": "./lib/hooks/register-tasks",
       "register-actions": "./lib/hooks/register-actions",
       "register-task-reporter": "./lib/hooks/register-task-reporters"
     }
