+++ packages/checkup-plugin-ember/src/hooks/register-tasks.ts
@@ -7,6 +7,7 @@ import EmberTestTypesTaskTask from '../tasks/ember-test-types-task';
 import EmberTypesTask from '../tasks/ember-types-task';
 import EmberTemplateLintDisableTask from '../tasks/ember-template-lint-disable-task';
 import EmberTemplateLintSummaryTask from '../tasks/ember-template-lint-summary-task';
+import EmberOctaneMigrationStatusTask from '../tasks/ember-octane-migration-status-task';
 
 const hook: Hook<RegisterTaskArgs> = async function ({ context, tasks }: RegisterTaskArgs) {
   let pluginName = getPluginName(__dirname);
@@ -17,6 +18,7 @@ const hook: Hook<RegisterTaskArgs> = async function ({ context, tasks }: Registe
   tasks.registerTask(new EmberTestTypesTaskTask(pluginName, context));
   tasks.registerTask(new EmberTemplateLintDisableTask(pluginName, context));
   tasks.registerTask(new EmberTemplateLintSummaryTask(pluginName, context));
+  tasks.registerTask(new EmberOctaneMigrationStatusTask(pluginName, context));
 };
 
 export default hook;
