+++ packages/checkup-plugin-javascript/src/hooks/register-tasks.ts
