+++ packages/cli/__tests__/commands/run-test.ts
@@ -14,9 +14,9 @@ import { runCommand } from '../../src/run-command';
 
 const TEST_TIMEOUT = 100000;
 
-class MigrationTask extends BaseTask implements Task {
-  taskName = 'migration';
-  taskDisplayName = 'Migration Task';
+class FooTask extends BaseTask implements Task {
+  taskName = 'foo';
+  taskDisplayName = 'Foo Task';
   category = 'fake1';
   group = 'group1';
 
@@ -28,9 +28,9 @@ class MigrationTask extends BaseTask implements Task {
   }
 }
 
-class MigrationItemCountTask extends BaseTask implements Task {
-  taskName = 'migration-item-count';
-  taskDisplayName = 'Migration Item Count Task';
+class FileCountTask extends BaseTask implements Task {
+  taskName = 'file-count';
+  taskDisplayName = 'File Count Task';
   category = 'fake2';
   group = 'group2';
 
@@ -71,7 +71,7 @@ describe('@checkup/cli', () => {
     it(
       'should output checkup result',
       async () => {
-        await runCommand(['migration', '--cwd', project.baseDir]);
+        await runCommand(['run', '--cwd', project.baseDir]);
 
         let output = stdout().trim().split('\n');
         output[7] = '<outputPath>';
@@ -96,10 +96,10 @@ describe('@checkup/cli', () => {
     it(
       'should output list of available tasks',
       async () => {
-        _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
-        _registerTaskForTesting(new MigrationTask(getTaskContext()));
+        _registerTaskForTesting(new FileCountTask(getTaskContext()));
+        _registerTaskForTesting(new FooTask(getTaskContext()));
 
-        await runCommand(['migration', '--cwd', project.baseDir, '--list-tasks']);
+        await runCommand(['run', '--cwd', project.baseDir, '--list-tasks']);
 
         expect(stdout()).toMatchSnapshot();
       },
@@ -109,7 +109,7 @@ describe('@checkup/cli', () => {
     it(
       'should output checkup result in verbose mode',
       async () => {
-        await runCommand(['migration', '--cwd', project.baseDir, '--verbose']);
+        await runCommand(['run', '--cwd', project.baseDir, '--verbose']);
 
         expect(stdout()).toMatchSnapshot();
       },
@@ -117,7 +117,7 @@ describe('@checkup/cli', () => {
     );
 
     it('should output checkup result in JSON', async () => {
-      await runCommand(['migration', '--format', 'json', '--cwd', project.baseDir]);
+      await runCommand(['run', '--format', 'json', '--cwd', project.baseDir]);
 
       let output = JSON.parse(trimCwd(stdout().trim(), project.baseDir)) as Log;
       expect(output).toMatchSnapshot({
@@ -131,7 +131,7 @@ describe('@checkup/cli', () => {
         let tmp = createTmpDir();
 
         await runCommand([
-          'migration',
+          'run',
           '--format',
           'json',
           `--output-file`,
@@ -151,47 +151,33 @@ describe('@checkup/cli', () => {
     );
 
     it('should run a single task if the tasks option is specified with a single task', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
 
-      await runCommand([
-        'migration',
-        '--task',
-        'fake/migration-item-count',
-        '--cwd',
-        project.baseDir,
-        '--verbose',
-      ]);
+      await runCommand(['run', '--task', 'fake/file-count', '--cwd', project.baseDir, '--verbose']);
 
       expect(stdout()).toMatchSnapshot();
     });
 
     it('should run with timing if CHECKUP_TIMING=1', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
 
       process.env.CHECKUP_TIMING = '1';
-      await runCommand([
-        'migration',
-        '--task',
-        'fake/migration-item-count',
-        '--cwd',
-        project.baseDir,
-        '--verbose',
-      ]);
+      await runCommand(['run', '--task', 'fake/file-count', '--cwd', project.baseDir, '--verbose']);
 
       expect(stdout()).toContain('Task Timings');
       process.env.CHECKUP_TIMING = undefined;
     });
 
     it('should run multiple tasks if the tasks option is specified with multiple tasks', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
-      _registerTaskForTesting(new MigrationTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
+      _registerTaskForTesting(new FooTask(getTaskContext()));
 
       await runCommand([
-        'migration',
+        'run',
         '--task',
-        'fake/migration-item-count',
+        'fake/file-count',
         '--task',
-        'fake/migration',
+        'fake/foo',
         '--cwd',
         project.baseDir,
         '--verbose',
@@ -201,20 +187,20 @@ describe('@checkup/cli', () => {
     });
 
     it('should run only one task if the category option is specified', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
-      _registerTaskForTesting(new MigrationTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
+      _registerTaskForTesting(new FooTask(getTaskContext()));
 
-      await runCommand(['migration', '--category', 'fake1', '--cwd', project.baseDir, '--verbose']);
+      await runCommand(['run', '--category', 'fake1', '--cwd', project.baseDir, '--verbose']);
 
       expect(stdout()).toMatchSnapshot();
     });
 
     it('should run multiple tasks if the category option is specified with multiple categories', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
-      _registerTaskForTesting(new MigrationTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
+      _registerTaskForTesting(new FooTask(getTaskContext()));
 
       await runCommand([
-        'migration',
+        'run',
         '--category',
         'fake1',
         '--category',
@@ -228,20 +214,20 @@ describe('@checkup/cli', () => {
     });
 
     it('should run only one task if the group option is specified', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
-      _registerTaskForTesting(new MigrationTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
+      _registerTaskForTesting(new FooTask(getTaskContext()));
 
-      await runCommand(['migration', '--group', 'group1', '--cwd', project.baseDir, '--verbose']);
+      await runCommand(['run', '--group', 'group1', '--cwd', project.baseDir, '--verbose']);
 
       expect(stdout()).toMatchSnapshot();
     });
 
     it('should run multiple tasks if the group option is specified with multiple groups', async () => {
-      _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
-      _registerTaskForTesting(new MigrationTask(getTaskContext()));
+      _registerTaskForTesting(new FileCountTask(getTaskContext()));
+      _registerTaskForTesting(new FooTask(getTaskContext()));
 
       await runCommand([
-        'migration',
+        'run',
         '--group',
         'group1',
         '--group',
@@ -257,15 +243,15 @@ describe('@checkup/cli', () => {
     it(
       'should run a task if its passed in via command line, even if it is turned "off" in config',
       async () => {
-        _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
+        _registerTaskForTesting(new FileCountTask(getTaskContext()));
 
-        project.addCheckupConfig({ tasks: { 'fake/migration-item-count': 'off' } });
+        project.addCheckupConfig({ tasks: { 'fake/file-count': 'off' } });
         project.writeSync();
 
         await runCommand([
-          'migration',
+          'run',
           '--task',
-          'fake/migration-item-count',
+          'fake/file-count',
           '--cwd',
           project.baseDir,
           '--verbose',
@@ -283,7 +269,7 @@ describe('@checkup/cli', () => {
       anotherProject.writeSync();
 
       await runCommand([
-        'migration',
+        'run',
         '--config',
         join(anotherProject.baseDir, '.checkuprc'),
         '--cwd',
@@ -298,7 +284,7 @@ describe('@checkup/cli', () => {
     it(
       'should run the tasks on the globs passed into checkup, if provided, instead of entire app',
       async () => {
-        _registerTaskForTesting(new MigrationItemCountTask(getTaskContext()));
+        _registerTaskForTesting(new FileCountTask(getTaskContext()));
 
         project.files = Object.assign(project.files, {
           foo: {
@@ -314,11 +300,11 @@ describe('@checkup/cli', () => {
 
         project.writeSync();
         await runCommand([
-          'migration',
+          'run',
           '**/*.hbs',
           '**baz/**',
           '--task',
-          'fake/migration-item-count',
+          'fake/file-count',
           '--cwd',
           project.baseDir,
           '--verbose',
@@ -328,7 +314,7 @@ describe('@checkup/cli', () => {
 
         clearStdout();
 
-        await runCommand(['migration', '--cwd', project.baseDir, '--verbose']);
+        await runCommand(['run', '--cwd', project.baseDir, '--verbose']);
         let unFilteredRun = stdout();
         expect(unFilteredRun).toMatchSnapshot();
 
@@ -343,7 +329,7 @@ describe('@checkup/cli', () => {
         project.addCheckupConfig({ excludePaths: ['**/*.hbs'] });
         project.writeSync();
 
-        await runCommand(['migration', '--cwd', project.baseDir, '--verbose']);
+        await runCommand(['run', '--cwd', project.baseDir, '--verbose']);
         let filteredRun = stdout();
         expect(filteredRun).toMatchSnapshot();
 
@@ -352,7 +338,7 @@ describe('@checkup/cli', () => {
         project.addCheckupConfig({ excludePaths: [] });
         project.writeSync();
 
-        await runCommand(['migration', '--cwd', project.baseDir, '--verbose']);
+        await runCommand(['run', '--cwd', project.baseDir, '--verbose']);
         let unFilteredRun = stdout();
         expect(unFilteredRun).toMatchSnapshot();
 
@@ -367,7 +353,7 @@ describe('@checkup/cli', () => {
       'should use the excludePaths provided by the command line',
       async () => {
         await runCommand([
-          'migration',
+          'run',
           '--cwd',
           project.baseDir,
           '--exclude-paths',
@@ -383,7 +369,7 @@ describe('@checkup/cli', () => {
         clearStdout();
 
         await runCommand([
-          'migration',
+          'run',
           '--cwd',
           project.baseDir,
           '--exclude-paths',
@@ -406,7 +392,7 @@ describe('@checkup/cli', () => {
         project.writeSync();
 
         await runCommand([
-          'migration',
+          'run',
           '--cwd',
           project.baseDir,
           '--exclude-paths',
@@ -420,7 +406,7 @@ describe('@checkup/cli', () => {
 
         clearStdout();
 
-        await runCommand(['migration', '--cwd', project.baseDir, '--verbose']);
+        await runCommand(['run', '--cwd', project.baseDir, '--verbose']);
 
         let hbsFilteredRun = stdout();
         expect(hbsFilteredRun).toMatchSnapshot();
