+++ packages/cli/src/index.ts
@@ -1,4 +1,4 @@
 export { run } from '@oclif/command';
 export { runCommand } from './run-command';
 export { getReporter } from './reporters/get-reporter';
-export { default as Checkup } from './commands/info';
+export { default as Checkup } from './commands/run';
